# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  template_internal_http_datatransfers_requests.ConnectIntegrationRequest:
    properties:
      access_token:
        maxLength: 4096
        type: string
      expires_at_ms:
        description: ExpiresAtMs нь токены дуусах epoch миллисекунд (0 бол хугацаагүй).
        type: integer
      provider:
        maxLength: 32
        type: string
      refresh_token:
        maxLength: 4096
        type: string
    required:
    - access_token
    - provider
    type: object
...
  template_internal_http_datatransfers_requests.GSpaceUploadRequest:
    properties:
      data:
        maxLength: 2900000
        type: string
      name:
        maxLength: 200
        type: string
    required:
    - data
    - name
    type: object
...
  template_internal_http_datatransfers_requests.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  template_internal_http_datatransfers_responses.EIDCertificate:
    properties:
      issuer:
        type: string
      key_type:
        type: string
      not_after:
        type: string
      not_before:
        type: string
      serial:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.EIDInfo:
    properties:
      certificate:
        $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDCertificate'
      civil_id:
        type: string
      document_number:
        type: string
      kyc_level:
        description: сертификатын түвшин
        type: string
      national_id:
        description: регистрийн дугаар
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.GoogleInfo:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      linked_at:
        type: string
      name:
        type: string
      picture:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.IntegrationResponse:
    properties:
      connected:
        type: boolean
      connected_at:
        type: string
      expires_at:
        type: string
      provider:
        type: string
    type: object
...
  template_internal_http_datatransfers_responses.UserResponse:
    properties:
      created_at:
        type: string
      eid:
        allOf:
        - $ref: '#/definitions/template_internal_http_datatransfers_responses.EIDInfo'
        description: |-
          EID нь eID-ээр нэвтэрсэн хэрэглэгчийн identity + сертификатын мэдээлэл.
          Нууц үгээр бүртгүүлсэн хэрэглэгчид nil (omitempty).
      eid_proxy:
        description: |-
          EIDProxy нь SSO eID proxy идэвхтэй эсэхийг заана — идэвхтэй бол иргэн
          локал eID linkage-гүй (SSO-ээр нэвтэрсэн) байсан ч eID PKI самбарыг SSO-
          гоор дамжуулан үзэж болно. Frontend eID хуудсуудыг үүгээр нээнэ.
        type: boolean
      email:
        type: string
      first_name:
        type: string
      first_name_en:
        type: string
      full_name:
        type: string
      full_name_en:
        type: string
      google:
        allOf:
        - $ref: '#/definitions/template_internal_http_datatransfers_responses.GoogleInfo'
        description: |-
          Google нь холбогдсон Google account-аас хадгалсан профайл. Google
          холбоогүй хэрэглэгчид nil (omitempty).
      id:
        type: string
      last_name:
        type: string
      last_name_en:
        type: string
      refresh_token:
        type: string
      role_id:
        type: integer
      token:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  template_internal_http_handlers_v1.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Gerege Template Platform V3.0
    url: https://github.com/snykk/go-rest-boilerplate
  description: chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture
    бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib
    Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/snykk/go-rest-boilerplate
  title: Gerege Template Platform V3.0 API
  version: "1.0"
paths:
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Өгөгдсөн refresh токеныг шалгаж, шинэ access+refresh хос үүсгэж,
        хуучин jti-г Redis-д хүчингүй болгоно. Аль хэдийн сэлгэгдсэн токеныг дахин
        тоглуулах нь 401 буцаана.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New token pair
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.UserResponse'
              type: object
        "401":
          description: Refresh token invalid, expired, or already revoked
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "403":
          description: Account no longer active
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: Refresh токеныг сэлгэж, шинэ хос буцаах
      tags:
      - auth
...
  /core/organizations:
    get:
      description: core.gerege.mn /api/organization/find — search_text (регистр г.м.).
        Хариуг дамжуулна.
      parameters:
      - description: regno / нэр
        in: query
        name: search_text
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: Gerege Core байгууллага хайх
      tags:
      - core
  /core/users:
    get:
      description: core.gerege.mn /api/user/find — search_text (core_id эсвэл регистр).
        Хариуг дамжуулна.
      parameters:
      - description: core_id / regno
        in: query
        name: search_text
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: Gerege Core хэрэглэгч хайх
      tags:
      - core
...
  /integrations:
    get:
      description: Тухайн хэрэглэгчийн холбосон провайдеруудыг буцаана (токенгүй —
        зөвхөн provider + хугацаа).
      produces:
      - application/json
      responses:
        "200":
          description: Connected providers
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/template_internal_http_datatransfers_responses.IntegrationResponse'
                  type: array
              type: object
        "401":
          description: Unauthenticated
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Холбосон интеграцийг жагсаах
      tags:
      - integrations
    post:
      consumes:
      - application/json
      description: Гуравдагч этгээдийн (Google Drive/Meet, Dropbox) OAuth токеныг
        тухайн хэрэглэгчид холбон шифрлэн хадгална. Frontend BFF нь token exchange-ийн
        дараа дуудна. Давхцвал шинэчилнэ.
      parameters:
      - description: Provider tokens
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.ConnectIntegrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Stored
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Malformed body / unknown provider
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "401":
          description: Unauthenticated
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Интеграцийн токеныг хадгалах
      tags:
      - integrations
...
  /users/me:
    get:
      description: Authorization header дахь JWT-ээс баталгаажуулагдсан хэрэглэгчийг
        уншиж, тохирох бичлэгийг буцаана (эхлээд in-memory кэш, олдоогүй үед Postgres).
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            allOf:
            - $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/template_internal_http_datatransfers_responses.UserResponse'
              type: object
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "404":
          description: User no longer exists
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Одоогийн хэрэглэгчийн профайлыг буцаах
      tags:
      - users
...
  /v1/gspace:
    delete:
      parameters:
      - description: Файлын нэр
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Gerege Space-с файл устгах
      tags:
      - gspace
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Gerege Space товч (файлууд + квот)
      tags:
      - gspace
  /v1/gspace/download:
    get:
      parameters:
      - description: Файлын нэр
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Gerege Space-с файл татах
      tags:
      - gspace
  /v1/gspace/upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: Файл (base64)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.GSpaceUploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Gerege Space-д файл оруулах
      tags:
      - gspace
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: /auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен
      (access token).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"