# internal/http/datatransfers/requests/requests_gspace.go
// Gerege Template Platform V3.0
// Gerege Systems Development Team болон Claude AI хамтран бүтээв, 2026.

package requests

// GSpaceUploadRequest нь Gerege Space-д файл оруулах body. Data нь base64 (≤~2MB).
type GSpaceUploadRequest struct {
	Name string `json:"name" validate:"required,max=200"`
	Data string `json:"data" validate:"required,base64,max=2900000"`
}